kagura: create thermanager.xml based on stock thermal config files
* Based on the actual values used by Sony
* Based on the actual sensors used by Sony

note: it looks quite sparse copared to previous platforms, but long
gone are the days we would measure individual cores. The high perf
cores of modern SoC means that huge, brief temperature swings on
each core are normal and correct.

Today we need to measure only the overall temperatures of each
component, which (for us) means the benefit of a nice, simple
thermal config.

Signed-off-by: Adam Farden <adam@farden.cz>